I08-1 Add way to display/hide ScriptFilesSection and PolarisationSection
When submitting a Polarisation scan, a BeforeScript will be added to the
ScriptFiles object in the MappingBean. But when changing to another
section, this script should not be set. Likewise, when running
polarisation scans, other Before/After scripts should not be set.

With this change,the dialog button is enabled and disabled and, a new
ScriptFiles object is created with null objects as Before and After
scripts every time the PolarisationSubmitScanSection is shown or is
hidden. And only the Polarisation Before Script is set when submitting
the scan.

Likewise, the PolarisationSection will be display/hidden depending on
whether this section is shown.

Change-Id: I18ef24ed36db055f44ee76662cc57e8092f48c42