B18-539 : Spreadsheet file reading performance improvements
- XML file list updating is now encapsulated in a new FileListWatcher class.
This maintains a list of XML files and their types as a map in
memory. A java.nio.file.WatchService object is used to monitor the directory
contents and trigger a map update if it changes.
- When changing the base XML directory, a modal dialog is displayed
to show the progress of the file list update process.
- Expensive file operations are now performed when necessary, rather than
every time a file combo box is selected in the spreadsheet table.
These changes are highly beneficial when dealing with directories
containing a large number (100s) of XML files.

Change-Id: I19308bfe0774dc85dae67f1133ff940680290518