I20-598 : Make XesSpectrometerScannable allowed angle range configurable
- The range of allowed angles when moving XesSpectrometerScannable objects
is now configurable. The initial limits are set to the same values as
previously (i.e. from MIN_THETA and MAX_THETA values in XesUtils).
- Angle validation is now carried out in XesSpectrometerScannable
(rather than in XesEnergyScannable) by using a PositionValidator.
- A method to return the allowed energy range has been added to
XesEnergyScannable and added to the IXesEnergyScannable interface.
This is used in client side code to ensure that energy range validation
in the GUI widgets is correct for each spectrometer object.

Change-Id: If627a1cd672472db4d57ed7feb801b8f982737dc