I08-1 Add functionality to run polarisation scans with energy scannable
This change contains the creation of a new section that extends
SubmitScanSection to set a before script and hides the scripts files
section. If a mapping section is set, it will add the parameters as an
environment of the script section.

Then, the PolarisationSubmitScanSection checks whether an energy model
was set. If it is not set, it will submit the scan with the before
script. If it is set, it will submit a first scan with the first energy
value and the before script. Then, the consequent scans will not contain
any before script and will have the energy values defined as beamline
configuration.

Change-Id: If586122864c6eb78b1e4061dd59953ed5ca1e3ad