DAQ-4808 : NexusDetector implementation to read images from Mjpeg camera stream
This is a new implementation of NexusDetector which returns the latest
image from an MJpeg stream from the 'readout()' method. This can be used
as an alternative to getting the detector image from the array plugin
(which may be problematic if the image size is large - due to max array
size limitations in the Epics IOC). It may also be used to capture
images from webcam streams - where there is no Epics control, and only
the stream is available.

Change-Id: I2b370bd92f26cfc98ec504a0a8950bd0955aa91f