set FilePathService in NexusScanDataWriterTestSetup
so that this service is set for all test classes that use that class
to setup their services. This is necessary now that we try to write
the visit id to the nexus file as experiment_identifier, as this
comes from that service. If the service is null, the test will still
pass, but somehow in Jenkins static state may be left over from previous
tests where the visit id is null, which causes the test to fail.

Change-Id: I42e1d4fbec9c97022a2d957e1e17877a827a1923