Add the option to show localStation log on an ssh shell connection
The GDA GUI client always displays the output of the most recent run of
localStation.py whenever the client is started. By setting
"gda.server.ssh.startupOutput = True" in common_instance_java.properties
this change allows you to get the same behaviour in the ssh client.

This is useful on beamlines which generate an "errors and warnings"
summary at the end of localStation.py, to avoid having to scroll back
through log messages to find the same information.

Change-Id: I3be7e5d3ad07be99403a732fbd2ddfe881bc6a22