[DAQ-4383] fix axes indices for NexusDetectorNexusDevice
The INexusTree returned by DummyNexusDetector should use 1-based
indices, not taking the scanRank into account. This is consistent with
real implementations of NexusDetector. When configuring the
NexusObjectWrapper for each axis field, NexusDetectorNexusDevice must
convert the index to a 0-based index, which takes the scan rank into
account (i.e. so that the index is of the actual dataset for the data).

Change-Id: Ie109dc3283530c84d6e49cae0645e1d52524ee97