Restore functionality of test in OsgiServiceBeanHandlerTest
This test did not do anything.

The stubbings were removed with the switch from initMocks to
MockitoJUnitRunner as they were unused. Note that the tests have since
been updated to JUnit 5 with Mockito extension.

The missing call to the processor has been added back so that the test
is working as originally intended.

Change-Id: I8f26da3bc42f3a1bd17bc387eaf1c83da2443d7a