Add a retryDeadband property to DummyMotor
If a scannable motor has returnDemandPosition=true, but is attached to a
DummyMotor, the logs are flooded with warnings about scannables
returning a "position based on real motor position rather than last
demanded position, as these differ by more than the configured demand
position tolerance (NaN)".

This change allows a non NaN value to be returned in dummy mode,
allowing the scannable to be defined in _common, while the referenced
motors are defined in live or dummy.

This also works as a demo of how to get example beans in javadoc to have
consistent indentation, where the indentation of the first line of the
bean is consistent with the indentation of the rest of the bean.

Change-Id: I5a90a4466639c5078f736ac772fefa9fd8aef20a