I14-1020 Add 'Calibrate' button to Tomography dialog
The button 'Calibrate' will calibrate the positions using the
TomographyFitSine method. The result for each axis (X, Y, Z) will be a
new TomographyCalibrationData and will be placed in the Jython namespace
and a script will be run printing out the fit for each axis. This can
easily be copied to other scripts that the beamline staff uses for
running Tomography scans.

Change-Id: Ib1be51f1a445ca97fefa75230483fcd953f8ca1e