Ensure the data file from a previous scan is closed before the next scan
Since there is no NexusTreeProviderProcessor.atCommandFailure(), the
open data file didn't get closed on failure.

This also meant that, as hdfFilePath is only set to null when the data
file is closed, and ensureFileOpen() short-cuts opening the file if
hdfFilePath is non-null, the path from any failed scan would be used for
all future scans.

Change-Id: I6b509dfdd9884f11d4ddb3a691f18c715ae67e40