Improve error reporting when client does not start
If there is a configuration error, it is unlikely that restarting the
client, (even with reset) is going to solve anything. For bean
exceptions stemming from loading the spring configuration, display a
more useful error pointing at an error in the configuration.

This also extracts the `showError` functionality into its own function
but the previous behaviour for unknown errors remains unchanged.

Change-Id: I7e8ad57bf30073072433c7c3e018cbe13e9c1b4f