Add SoftwareStartStopHDFWriter to control file writer outside of scan
This purely configures a NDFileHDF5 object, so it is assumed that the
detector is already configured and acquiring otherwise. The file writer
plugin is configured as follows:
 - Uses SWMR
 - Sets the directory passed to #start(String) after creating the
   path if it doesn't exist
 - Sets file prefix as passed to SoftwareStartStopHDFWriter
 - Turns on auto increment
 - Sets file template as "%s%s%d.hdf5": {dir}/{prefix}{number}.hdf
 - Sets write mode to Stream and number of points to capture to 0,
   meaning don't stop (until manually stopped)

Because GDA does not set the full file name, there is a slight race
condition where the full file name is blank when starting the writer.
Therefore we sleep until that PV is not blank before returning from
start.

Change-Id: I44a32c63e2534174db1ce8cfe8848c7af8387fb3