k11: Add EnvironmentDataWriterController which responds to experiment
This object is configured with an environment data writer and listens to
experiment events to start and stop the writer accordingly. It also
publishes the environment data file name so that the experiment control-
ler can link to it from the main experiment file.

To support this, ExperimentEvent now includes the experiment file URL.

Change-Id: Ia73e425dc9f9454bf41c287d8f7dbe94ce49fd96