DAQ-4926 Add FAILED_REQUEST Status
- Add a new Status called FAILED_REQUEST: The aim is to identify the
beans that have been unsuccessfully aborted, where an exception was
thrown when trying to terminate. When this happens, the scan gets stuck
in REQUEST_TERMINATED. This scan cannot be called to terminate again
and, clearing the queue doesn't help as only the active scans are called
to be terminated again. The solution in the past has been to restart gda
servers.

With the new Status, when an exception is thrown while trying to abort
all the scan components, the bean status is set to FAILED_REQUEST. Then,
when clearing the queue, all the active scans and the ones with the
FAILED_REQUEST status will be cleared, so other jobs can start running.

Change-Id: I28b3fe259cf54bb40446a14033484476b002c040