Add abort support to bluesky controller
Blueapi already supports aborting running plans, accessing this from GDA
is the most recently requested feature. It is particularly useful if the
blueapi worker becomes frozen and unresponsive.

In this change:
- Add awareness of blueapi's abort interface to the bluesky controller.
- Add a simple "abort" method which both commands the worker to abort
and waits for the abort to complete, on a timeout.
- Create a Jython function wrapping the above method.
- Send an abort command if running "run_plan()" in Jython and using the
Jython command interrupt (CTRL-C in the remote shell or the cancel
button in the GDA client)

Change-Id: I3a3cc2a2b49e3d8d3a83f4b533bd88ea2d5ef24c