I20-611 : Allow XES epectrometer detector demand positions to be rounded
An optional array of detector demand precision values can now be set on
XesSpectrometerScannable. This allows the x, y, pitch values for the
detector motors to to be rounded before they are sent to Epics (if a
precision is set to zero, then no rounding is carried out).
Unit tests have been updated to check that rounding is applied
correctly.

Change-Id: I20f7e6f67ff7574a645d9fa2fea0b50f2b39d5e7