B18 : Make QexafsTestingScannable delegate motion to a scannable
Instead of QexafsTestingScannable being a scannable motor which moves
during a continuous scan, it now delegates all motion related operations
to another scannable.
This makes it easy to drive different motors in continuous scans, by
simply changing the delegate scannable.
It also makes it possible to use QexafsTestingScannable in continuous
scans to control objects that are not ScannableMotors.

Change-Id: Ib0ab369a611488921127f789cdbaff1ec4f79281