Add TokenStreamHighlighter for JythonShell
Use the same tokeniser as the translator to provide basic syntax
highlighting. On the downside it is not able to differentiate between
as many highlighting classes (method name vs parameter name etc), but
on the upside there is no instance of Jython required to run pygments.

For now only a single, basic theme is provided but custom ones can be
provided by the user (and/or added to the 'themes' file).

Change-Id: I3eb9eb5ddfafeaa6caa35eeef2320de3b5e4544e