B18-561 : Updates for Xspress4
- A sleep has been added after starting hdf and meta writers to allow Epics
PVs time to update, and help avoid race conditions. The sleep duration can be
controlled setting hdfWriterWaitTimeMs on the Xspress4Controller (default
is 500ms).
- Detector 'start' and 'stop' methods have been added to Xspress4Detector.
These start and stop the detector and hdf writers (including Odin metawriter).
The stop method optionally blocks until the detector is in an idle/disarmed state.
- The detector is explicitly stopped at start of scans to ensure it is Idle
before attempting to configure it.
- Xspress4 buffered detector now prepares the detector correctly
using setContinuousMode function. Previously, the atScanStart function was
called first with the incorrect number of frames, followed by setContinuousMode
with the correct (larger) number. This put the detector in an inconsistent
state, causing it to miss triggers for frames at the end of many Qexafs scans.
- Xspress4 buffered detector now filters out NaN and infinite values from
DTC factor data and replaces them with 1.0. A warning is logged if this
occurs.

Change-Id: I91478afdf34e1a065bd76200e993781a84554392