[DAQ-4932] Fix error finding PreprocessorService
This error was causing scans to fail to run. The problem seemed to be
that the component declaration file didn't specify the class/interface
that the PreprocessorService class implemented. This has been fixed
by adding an interface (which may not have been strictly necessary)
and declaring that as the provided service interface.

Change-Id: I4783ef3d126dd00983974b1056e1b6f53e9fd417