I16-650 (9.33) Add support for ScannableComponentMetadataAttribute
This is analogous to ScannableComponentField but for attributes rather
than fields.

Since there is no SpringExpressionAttibute, the only way to populate
attributes dynamically was to create a single scannable per attribute.

This problem was solved for fields by ScannableComponentField which
allows a single scannable to provide values for multiple fields, keeping
related fields together. ScannableComponentMetadataAttribute likewise
allows a scannable to provide attributes too. This means a single
scannable can provide all values needed for both attributes and fields.

ScannableComponentMetadataAttribute includes better support for Spring
Beans, including an afterPropertiesSet check that scannableName and
either componentName or componentIndex is set.

It also includes the same code copied from AbstractScannableNexusDevice
with the same TODO that it should eventually be refactored.

Change-Id: I36379e8aaeb25c2fbd21fc868a1cea237e58379d