DAQ-3621 Add interactive history search to JythonTerminalView
This is intended to be similar to the reverse history search available
in readline (via Ctrl-R). Ctrl-R will open a history popup filtered on
any existing input. Typing while the popup is open will filter the
history items displayed. The chosen history item will replace the
command when selected by hitting enter.

'Normal' autocompletion is available when filtering history although it
is better to confirm a selection via either tab or space as using enter
will select both the completion and the first matching history item.

Change-Id: I28014449b085ff2988537341582650128fc30f88