k11: Disable selecting processing file once it's set
This is because leaving it enabled suggests it can be edited but editing
is not implemented. Instead, selecting a different file creates a new
processing entry... Disabling instead of implementing since this part
will soon be replaced with the processing UI used in Mapping.

Also refactored the handle method for file-based processing options to
reject a processing entry if the user dismisses the file selection
dialog.

Finally moved some @Autowired annotations from fields to constructors
to simplify testing

Change-Id: I91f88ad8b61e9e030cfded76aa5c14c539c73666