reinstate ScanBase.sortArguments as deprecated method
This public static method was renamed without the original method.
This will break jython code that calls the old method. The
original method should have been kept as a deprecated method that
delegates to the new method.

This change reinstates the original method, marked as deprecated, as a
delegate to the renamed method.

Change-Id: Ia493aaaae0beb49555d437d21b55152a635cab07