[DAQ-2239] handle duplicate aliases for Finder singleton methods
previously if the same instance appeared with multiple aliases this
caused an exception to be thrown indicating that there are multiple
instances of the singleton class. The same instance can sometimes have
multiple aliases (two in the case found) in the Finder due to a
possible bug in the spring handling.

Change-Id: I53ff026d4f0d85ea3ad72a399b5f9a4df25da330