K11-1398 Fix NPE in LatestSwmrFinder
A file's tree was optimistically parsed as soon as #fileLoaded was in-
voked. There is no guarantee, however, that the tree has been loaded at
this point. The parsing in now extracted from the function and only
called when the tree isn't null; otherwise it is postponed to the
following #refreshRequest invocation.

Additionally, the refresh and tree reading is now performed in a back-
ground thread to avoid blocking the subscription thread.

Change-Id: I8d876060629c4101c8e77d30f0b619760cdd7692