I14 Add Malcolm monitor to pause/resume mapping queue
Implementing a new Malcolm monitor that observes a PV that represents
the state of Malcolm, whether it is Healthy or there has been a failure
and will proceed to restart.
The aim of this implementation is to pause the queue while Malcolm is
being restarted. So, if there is a failure that can be solved by
restarting Malcolm, some scans that were queued do not go straight into
FAILED status.
Change-Id: I920687be44e1a58f7a672ed75ee0b84df007ab32

Change-Id: I28b9cd8a08001e0f4a0914f9e0dc6b342863916b