Add spring parser to let relative paths be used as property values
When the location of the spring XML is not known, it is difficult for
values used in the beans to reference relative file paths. This parser
creates a String bean containing the resolved relative path that can be
used as any other property. Spring's normal type coercion can still be
used (eg to set a File or Path field using the relative path).

    <property name= "path">
        <gda:relative-path path="../../scripts" />
    </property>

Change-Id: I19228d95ab5e24fd1bbcce0f807a50af057f6b19