I18-610 Add methods to find the last good point in Malcolm devices
When GDA pauses a Malcolm device, it sends a lastGoodStep attribute
which defines the point that Malcolm should seek when it's called to
resume.

This number is usually -1, which means that it should go back to the
previous step. However, due to the way Odin devices write their data,
this will be different for each Odin Malcolm device. The last good point
will be calculated based on different variable parameters like chunk
size, number of written frames, etc.

The first device that will be tested for this pausing/resuming behaviour
will be the Xspress3 Odin, but this can be extended to other Odin
devices.

In this change, I defined a new optional field in the Malcolm Model that
corresponds to the SeekStrategy implementation that will be called when
pause() is called on a PausingOdinMalcolmDevice, to calculate the last
good point and send this message to Malcolm.

Change-Id: I3e42146b6ee33dfa5fba456a57d6a0d59d180749