I07-610 Ensure that jython syntaxerrors are reported
If a syntax error is raised from a command containing a non-ascii
character, a Jython bug means the exception is not raised correctly and
an IllegalArgumentException is thrown instead. This would not previously
be reported to the user via the Jython terminal and the command would
appear to be successful.

Until we can update to a version of Jython where this is fixed, print
any exceptions raised to the Jython terminal.

Change-Id: I6256b0266d7224de4e7a5ff850e783dcc25673bb