DAQ-3775 Add TemporaryContextClassLoader to handle TCCL
There are several uses of the pattern of saving the current thread's
context class loader, setting the CCL to the class loader provided by
the GDAClassLoaderService, running a block of code and reverting the CCL
back to the original in a `finally` block.

    var previous = Thread.currentThread().getContextClassLoader();
    try {
        var newClassLoader = getClassLoaderService().getClassLoader();
        Thread.currentThread().setContextClassLoader(newClassLoader);
        // Code requiring new class loader
    } finally {
        Thread.currentThread().setContextClassLoader(previous);
    }

This change adds a TemporaryContextClassLoader class that can be used in
try-with-resources blocks and adds a factory method to the service to
create instance of it wrapping the service's class loader. These can
reduce the above to

    try (var tccl = temporaryClassLoader()) {
        // Code requiring new class loader
    }

Change-Id: I6c3470b7c6f205bc64af8a0978e65abdfb8c4ee6