I09-530 - Improve sequence file loading to handle NPE
If an error occurred while opening a file (for example file doesn't
exist or incompatible file format), then a NPE would be thrown and the
GUI would not be able to load a new sequence file as it would forever
run into this exception. It now checks for this and also alerts the user
if an issue occurred when opening a file.

Change-Id: I4eed7dc9c7fa3368f0fa9f6bd18e5f50efca09c4