K11: Fix filter box not matching saved scans
The text in the filter box was not matching the names of scans! This is
because PatternFilter#isLeafMatch takes the label provider of the viewer
and filters its #getText(element) value. In the case of TreeViewer-
Builder, no label provider is set because TableViewerColumns are created
instead, each with their own label provider. When you don't set a label
provider, getLabelProvider() returns a LabelProvider instance which
implements getText(element) as element.toString().

The solution in this change is to set the first column's label provider
as the viewers too.

Change-Id: I62bee4537ed1784048c0453bbeb47c0720f58229