B18-539 : Spreadsheet view updates for using new detector parameters
- SpreadsheetViewHelperClasses#generateNewScans now correctly creates
copies of the detector XML files when using new detector parameters.
- DetectorConfig parameters in the spreadsheet view now show the name of
the config file rather than the full path (all config files are in the
same directory as the other parameter XML files, so the full path is not
needed).
- Bugfix update to ensure that font generated in CheckboxLabelProvider
is disposed of correctly when the label provider is disposed.

Change-Id: Iaaf5568fd4c6d576d4cb496d4af9cea73550ef5f