I09-541 - group together commands to make sequence editor undo / redo more consistent
When updating a field in the UI, this can in turn update many other
fields e.g spectrum energy range width would update centre, low and
high. When pressing undo / redo in the sequence editor, this would undo
/ redo each of these fields one by one making it possible to get a
region into an inconsistent state. This change makes it possible to
group many commands together to execute at the same time. This means
when a undo / redo are called, it calls the undo for this group at the
same time so it can no longer be in a inconsistent state.

Change-Id: Id99f65a4a9f040659965b0792046aeb2a3b7805d