I20-632 : Use correct limits in XES 'simulated positions' view
The XES 'Simulated positions' GUI now uses limits from Bragg scannable
to determine the allowed range of Bragg angles and Energy that can be
achieved using the controls. A dialog box is displayed if any move
outside the angle or energy limit is attempted.
The GUI limits are set from the angle range of the XESBragg scannable
when the view is created. So, if the Bragg angle range or crystal cuts
of the underlying XESEnergy/Bragg scannables change, the view should be
re-opened to get the updated limits.

Change-Id: I1899d08482e254e29a634ffb048b0c5385f3b5a3