BufferedScaler : Add 'manual start' option
When using internal time frames in BufferedScaler, instead of starting
immediately when setContinuousMode is called, the time frames can be set
and the Tfg not started. The tfg can be started later on when needed by
calling the startTfg() method. This behaviour can be controlled
using setManualStart(true/false) as required.

Change-Id: Id98efe30c2d15e9a512533833efb63822bbcaf59