I18 : Xmap updates
Updates for NexusXmap :
- A sleep can be added before reading out data in collectData. This
gives the Epics record chance to update after each frame has been
collected before reading out the values.
- Detector can be run in 'Hardware triggered' collection mode. In this
mode the detector is armed at the start of the scan and stays armed for
the duration. Collection of each frame of data is is triggered by
external TTL trigger.

Minor update for EpicsXmapController :
- Cache numberOfBins in EpicsXmapController so that getNumberOfBins()
does not need to use channel access each time it is called. (This works
around an intermittent channel access failure for this PV which can
cause scan to fail).

Change-Id: Ia0ce7db692ca0831f9c1b9b60265762d3f4295db