I18 : LiveStreamViewRoiControls for setting area detector ROI
This class allows the area detector ROI to be set and retrieved from
a live stream view. It adds two button controls : one to set the
detector ROI from the ROI currently set in the plot window, and one to
set the plot window ROI from the ROI currently set in the area detector.

Change-Id: I95b355062f4982315bacc065726d47b61232f9f6