Set java.awt.headless mode in server product file
This property should be set for all servers. When run in live mode it is
set by the launcher (and may default to true if the JVM detects that
there's no DISPLAY), but for IDE launches it should be set to ensure the
environment is consistent.

Setting it in the product file prevents each config having to set the
same property.

Change-Id: I09c3288de035755ed3b164c3e1e3392ccabfff93