I14-1066 Record SampleZ position based on sample stage rotation
The value of SampleZ depends on the value of stage1_rotation. Beamline staff needs to know both values when running the tomography script. For instance, if stage1_rotation is 3.45, SampleZ is 7.3, if stage1_rotation is 4.56, SampleZ is 12.34 (made up values).

There is a text box displaying the value of stage1_rotation, and when a user presses "Record position", the corresponding value of SampleZ will be displayed. Then both values can be passed on to the script once the tomography script is run.

If stage1_rotation changes its position after the "Record position" button is pressed, the SampleZ text box will be blank and the user will have to press the button again to record the position.

Change-Id: If1faa7581865e95a5800e443ac2b4f90889e5a3d