I14-1066 Add timestamp to calibrated data
Tomography data is calibrated through the TomographyConfigurationDialog. The values are passed to a Jython script when submitting the tomography scan. However, the user does not have any indication of which data is being passed. Because the values only hold meaningful information for the script but not for the user, I have incorporated a label with a timestamp of the time when the data was calibrated.

Also:
- Refactor some code (changed a class for a Record).
- Add dispose method


Change-Id: Ie4186b4f8596a655362c9c63c8d86e066123e771