DAQ-3829 Update scan/script controls when client starts
If a scan/script script is already paused when a client is started, the
controls would not reflect to correct state. Ensure the state is correct
by querying the server state at startup.

There is a race condition here if the state changes between querying the
server and the listener being added but this is still more likely to be
correct than the previous version.

Change-Id: Idf7ce9182628cb85254aa8ebcb261afcdf53db65