DAQ-5167 Restore string array property handling
Prior to upgrading the configuration library as part of the config
loading redesign, multiple values could be set for a property by setting
the value to a comma-separated string. This is not supported by default
in the new version of the commons configuration and needs to be manually
enabled.

If multiple values are passed to repeated keys, the arrays of each are
concatenated into a single array containing all values.

Change-Id: Icc9fa84e9f51c81a40ed33c86267d54421599c79