I18-634 : Add functionality to BufferedScaler for laser experiment
Add extra options to BufferedScaler to allow more fine-grained control
over the command sent to the Tfg :
- External start with TTL triggers can be switched on and off
- Time frames used for data collection be externally or internally
triggered.
- A custom line can be added to the start of the timing group. This can
be used to implement additional functionality when required (e.g.
trigger another device, wait for a trigger to help with sychronisation
etc).
- The 'getNumberFrames' method now accounts for cycles - returning 0
until all cycles have completed, with readout of all frames once the Tfg
is in an Idle state.
- Unit tests have been updated to test the new options.

Change-Id: I04c99c995d5ae04d16d48ad07c25eb3c9a4003de