I08-401 Sample Transfer System
SampleTransfer handles the execution of a sequence with a number of
steps as defined in the XML configuration.

The sequence will contain server-side actions and steps that require
user confirmation to continue.

The execution flow is managed by the sequence commands: START, RESUME,
STOP and SHUTDOWN.

The dialogue also includes a view of the cameras.

Change-Id: Ied414ec471c7bfe3b21b6bda224e9638d5ee0aee