B21-267 Make Eiger timeout configurable
Replace hard coded 3 second timeouts and static ODIN_TIMEOUT value in
NcdEpicsEiger with an eiger_timeout property.

On B21, the expectedTriggers.caput was observed to take 5 seconds, which
is definitely longer than it's 3s timeout, and marginal with respect to
the 5s ODIN_TIMEOUT value. Make this configurable so that beamlines can
specify an unusually long timeout if required to get the beamline
running reliably again before the underlying problem is solved.

Change-Id: I6ad97bd7dcb7374146c67aa2c14211afaa08dbc4