Improve gda_completer's handling of classes
If a Jython class extends a Java class, Jython converts `setX`/`getX`
methods into attributes that can be accessed directly as `x`.

When calling `dir()` on one of these classes, these generated attributes
are returned even if they can't be accessed on the type directly
(requiring an instance of the type). This catches the exception raised
when the attribute is accessed (to test for function/attribute
metadata), and excludes the option from the generated list to as users
would not be able to access it either.

Change-Id: Icf239d72a9389d9a5cdc3a64d337273561b68ccf