[DAQ-5050] add property to NexusMetadataDevice to set nexus object name
within parent group, instead of using the name of the nexus device.

Since INexusDevices are registered with the INexusDeviceService by name,
the names has to be unique. If the same name is always used for the
nexus object as for the nexus device, it is not possible to add
nexus groups with the same name to different parent groups using
INexusDevices registered with the INexusDeviceService for both.

Change-Id: I5bb7b6b56709c032130723c7f8333ee0144f77df