I20-635 : Add checks to demand positions when moving XES spectrometer
Checks have been added when moving XES spectrometer to ensure that NaN
values are handled correctly. A device exception is now thrown if a non
finite number is used for the energy, Bragg angle or motor positions.
This ensures that NaN is never passed to Epics as the demand position
(to avoid unexpected behaviour).

Change-Id: I0ba4161c2437755fdeded0c2a5247d43151a41f3