I09-551 Improve sequence file loading
If a feature is not found in a sequence file, it would throw an error
and prevent the user from opening the file. This will now attempt to
recover and load the file again as the emf framework auto removes the
invalid features.

Additional changes
- Refactor initialisation and refreshTable functions. Initialisation is
only called once, so does any setup the view needs. Refresh table is
only function that loads in sequence files now. Initialisation before
was a hybrid of the two.
- Fixed Creator view not loading in excitation energies correctly.
- Fixed element set saving error message to file and stopping the
monitor. Setting elementSet is now done by single function.
- Disposed of commands that were executed to prevent memory leaks.
- Added additional helper function isSingleSource() to
RegionDefinitionResourceUtil.

Change-Id: I2fb5e8f6480c1515b48056e72b7d98670833457c