I09-526 - Only dirty sequence file when excitation energy is selected and not from scannable move.
- Dirtying of file from energy change now only applies when energy
changes from hard to soft or soft to hard. This way makes it so file
format change is not necessary as server only needs to know which
shutter to open anyway.
- Updated scannable motor listeners. Removed threads constantly checking
if energy scannables are busy. Instead, thread spawns when update is
received and is similar to live controls implementation.
- Removed unused code. Changed unneeded protected functions and
attributes to private.
- Made RegionViewCreator more generic for RegionViewLive to reuse.
- Added region name to RegionValidator log

Change-Id: I5625dca12211ff56df6fb3c94e4444628a17be36