I09 - Improve region validator logging
Several lines of logging was done for each region validation. I have
refactored this class to be more logical and clear while also changing
the logging to be a single line for each region validation while
combining the error message given to the UI to be logged as the single
message. This contains all of the values used to get the validation
value. A valid region is also logged with this information (as debug).

Change-Id: Ie48c75d85b5fadab3801be9c86a51a571a658fd2