[DAQ-5104] - Correct error message of no registered device or scannable being logged when one was present.
For the fix of prioritising registered nexus devices over scannables,
the error message was logged that no registered device or scannable was
found when it had only checked for a registered device. I've brought
this error message outside of the getRegisteredNexusDevice methid and
logged it if the scannable and registered ones are not present instead.

Change-Id: I930a19eb67c03055b8299673154d3d9b3d8fa82c