I18-638 : Allow mapping background stream to work when zoom level level cannot be determined
This improvement allows the mapping background stream to continue to work
even if the zoom level cannot be determined (e.g. due to the VMA zoom
PV being unavailable due to crashed IOC). In this case, a warning
dialog is shown to the user and image pixels are used for the axis values.

Change-Id: I4c7d19dedddad7a2835459387c2685a3a32d603b