I09 - Remove unnecessary setRegion calls
In SequenceViewLive, it would receive and handle updates from the GDA
class EW4000 to update the region status. However, it also set region
status directly from epics. I have removed this listener from epics to
stop it being set from multiple places and wasn't needed because if
triggered via epics, it may not be the correct current region to update
the status.

This also simplifies a lot of UI code to be shorter and more concise.

Additional bug fixes:
- Fixed time remaining calculation.
- Fixed excitation energy not taking into account canEdit in
RegionViewCreator
- Fixed RegionProgressComposite number of iterations being one less than
it should of been.
- Fixed channels not being destroyed on dispose method

Change-Id: Ib29d1401808cc87cb50b8d6949cfd2b6e23bc144