I09-579 - Change ProcStartStop to create monitor and monitorChannels to collectData
For I09, they have a camera, nixswr, which freezes when used with the
pos command. NXDetector does not support the use of the pos command as
shown via UNSUPPORTED_PART_OF_SCANNABLE_INTERFACE, however it will still
attempt to collect data and then throw an error after when trying to get
the position. This camera freezes with pos because the epics layer
callbacks are set via atScanStart method which is never called with pos
and therefore never sets up the collection strategy decorators.

Changing the ProcStartStep to create the monitorChannel and monitor
should fix this as the channels will always be created. The NXDetector
will then fail naturally with the pos command via
NXDetector.getPosition()

Change-Id: I78b47d416b2c98a6385dd8769db9e6ed9131d720