I09-582 - Prevent client changing region status from pgmenergy change event in the middle of a scan
If pgmenergy or dcmenergy are moved externally from the scan (or send
notifications because pgmenergy fluctuates around a point), it will
trigger a region validation and resets the status of the region even if
they are already completed or currently running. This fixes this by
preventing status changes during a scan.

Also fixes issue with SES Creator where pgmenergy wouldn't correctly set
user input.

Change-Id: I752082eeaa21190dda86bf985cd763ff205c2297