I18-642 : Add tomography scan to mapping GUI
This change is to add tomography scan to the mapping GUI. When I18SubmitScanSection
is shown in the GUI (e.g. via SubmitScanSelector) it automatically sets
the fast and slow axes to t1x and t1theta respectively. The previous fast
and slow axes settings are restored when the view is hidden.

Change-Id: I94a0ce3ebd23a012b0def9edca80345a9c68b370