I20-646 : Ensure XES positions are consistent with radius and crystal cut
- XesSpectrometerScannable now updates the radius value from radius
scannable before moving to a new position, and before calculating all
the motor positions. This ensures calculated positions are consistent
with current radius scannable value.
- The XES simulated positions GUI now listens to the crystal cut, radius
and material scannables, and immediately recalculates the positions
if any of them change. If this occurs, the energy position and limits
are also updated from the current Bragg angle.

Change-Id: Ie73b99a3d70950722151a682ee59ca0df0fb9f90