DAQ-5228 - Filter out duplicate detector names from NexusScanDataWriter when creating per scan devices
Discovered by accident when investigating DAQ-5237, but if you add a
detector as a per scan monitor and the detector also takes part in the
scan, then you end up with duplicate scan roles from the check I added
in NexusScanInfo. This fixes it by removing the duplicate scan roles by
removing the device being created as a per scan as taking part in the
scan takes priority. Same as I did to the scannables.

This is also mirroring the behaviour of NexusDataWriter.

Change-Id: I72178de08e448224e79108ac05abff72da3ae6a1