K11-2200 Aggressively ask for sample name
A sample name (backed by a PV) must exist before a scan is saved
or submitted. If it does not, the client will produce a popup prompting
for a name to be provided.

New Spring component PVSampleNameProvider uses property "sample.name.pv"
to determine the PV to read.

ScanRequestFactory now adds sample metadata.

Change-Id: Id8387948be794646a3eca4f67e317fce483eacd8