DAQ-5241- Fix NexusMetadataUtility to support configured scannables without a nexus device
You can add scannable names to commonBeamlineDeviceConfiguration and the
framework will automatically wrap them with a nexus device when running
a scan. However, NexusMetadataUtility crashes because it cannot find a
device with the name when doing meta.ll(). This changes makes it so if a
scannable is found and it cannot find INexusDevice, it will
automatically wrap it with a NexusMetadataAppender.

Change-Id: Ibfd722325fa188b8e280783f7a1fc324c379ae70